requirements.txt
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
fake-useragent==1.5.1
idna==3.7
importlib_resources==6.4.0
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
urllib3==2.2.2
zipp==3.19.2
